Add InitService
Added an InitService which runs when the AppComponent initializes in
order to send a GET request to the backend to acquire an XSRF token.
This is because when a user wants to register as his first thing
navigating to the page there won't be any XSRF tokens present, as such a
POST to /register will fail because of the backends CSRF protection.

Added, changed some minor stuff here and there.